+++ example/lib/menus.dart
@@ -68,6 +68,7 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                       ('Blue', Colors.blue),
                     ],
                     popupProps: PopupProps.menu(
+                      menuProps: MenuProps(align: MenuAlign.bottomCenter),
                       fit: FlexFit.loose,
                       itemBuilder: (context, item, isSelected) => Padding(
                         padding: const EdgeInsets.all(8.0),
@@ -97,7 +98,7 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                       }
                       return RotationTransition(
                         turns: AlwaysStoppedAnimation(r / 360),
-                        child: Image.asset('networks.png', height: 164, width: 164),
+                        child: Image.asset('assets/images/networks.png', height: 164, width: 164),
                       );
                     },
                     clickProps: ClickProps(borderRadius: BorderRadius.all(Radius.circular(50))),
@@ -178,7 +179,7 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
               Padding(padding: EdgeInsets.all(8)),
 
               Container(
-                height: 380,
+                height: 500,
                 padding: EdgeInsets.symmetric(vertical: 18, horizontal: 8),
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(16),
@@ -188,7 +189,6 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                   ),
                 ),
                 child: Column(
-                  //mainAxisAlignment: MainAxisAlignment.spaceBetween,
                   children: [
                     Text("Example for customized menu"),
                     Padding(padding: EdgeInsets.all(8)),
@@ -202,6 +202,7 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                       ],
                       dropdownDecoratorProps: DropDownDecoratorProps(
                         dropdownSearchDecoration: InputDecoration(
+                          contentPadding: EdgeInsets.symmetric(vertical: 6),
                           filled: true,
                           fillColor: Color(0xFF1eb98f),
                           border: OutlineInputBorder(
@@ -220,22 +221,21 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                       ),
                       dropdownBuilder: (context, selectedItem) {
                         return ListTile(
-                          dense: true,
                           leading: Icon(selectedItem!.$1, color: Colors.white),
                           title: Text(
                             selectedItem.$2,
-                            style: TextStyle(color: Colors.white, fontSize: 16, fontWeight: FontWeight.bold),
+                            style: TextStyle(color: Colors.white, fontSize: 18, fontWeight: FontWeight.bold),
                           ),
                         );
                       },
                       popupProps: PopupProps.menu(
                         itemBuilder: (context, item, isSelected) {
                           return ListTile(
-                            contentPadding: EdgeInsets.symmetric(vertical: 10, horizontal: 12),
+                            contentPadding: EdgeInsets.symmetric(vertical: 8, horizontal: 12),
                             leading: Icon(item.$1, color: Colors.white),
                             title: Text(
                               item.$2,
-                              style: TextStyle(color: Colors.white, fontSize: 16, fontWeight: FontWeight.bold),
+                              style: TextStyle(color: Colors.white, fontSize: 18, fontWeight: FontWeight.bold),
                             ),
                           );
                         },
@@ -261,9 +261,9 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                               Flexible(
                                 child: Container(
                                   decoration: BoxDecoration(
-                                      color: Color(0xFF1eb98f),
-                                      shape: BoxShape.rectangle,
-                                      borderRadius: BorderRadius.circular(8),
+                                    color: Color(0xFF1eb98f),
+                                    shape: BoxShape.rectangle,
+                                    borderRadius: BorderRadius.circular(8),
                                   ),
                                   child: popupWidget,
                                 ),
@@ -321,7 +321,44 @@ class _MenuExamplesPageState extends State<MenuExamplesPage> {
                           ),
                         ),
                       ),
-                    )
+                    ),
+                    Padding(padding: EdgeInsets.only(top: 32)),
+                    DropdownSearch<String>(
+                      items: (filter, loadProps) => ["Item 1", "Item 2", "Item 3", "Item 4"],
+                      dropdownDecoratorProps: DropDownDecoratorProps(
+                        dropdownSearchDecoration: InputDecoration(labelText: 'Top Left Menu', border: OutlineInputBorder()),
+                      ),
+                      popupProps: PopupProps.menu(
+                        fit: FlexFit.loose,
+                        constraints: BoxConstraints(maxWidth: 200, maxHeight: 100),
+                        menuProps: MenuProps(align: MenuAlign.topStart),
+                      ),
+                    ),
+                    Padding(padding: EdgeInsets.symmetric(vertical: 8)),
+                    DropdownSearch<String>(
+                      dropdownDecoratorProps: DropDownDecoratorProps(
+                        dropdownSearchDecoration:
+                            InputDecoration(labelText: 'Bottom Center Menu', border: OutlineInputBorder()),
+                      ),
+                      items: (filter, loadProps) => ["Item 1", "Item 2", "Item 3", "Item 4"],
+                      popupProps: PopupProps.menu(
+                        fit: FlexFit.loose,
+                        //constraints: BoxConstraints(maxWidth: 200, maxHeight: 300),
+                        menuProps: MenuProps(align: MenuAlign.topCenter),
+                      ),
+                    ),
+                    Padding(padding: EdgeInsets.symmetric(vertical: 8)),
+                    DropdownSearch<String>(
+                      dropdownDecoratorProps: DropDownDecoratorProps(
+                        dropdownSearchDecoration: InputDecoration(labelText: 'Top Right Menu', border: OutlineInputBorder()),
+                      ),
+                      items: (filter, loadProps) => ["Item 1", "Item 2", "Item 3", "Item 4"],
+                      popupProps: PopupProps.menu(
+                        fit: FlexFit.loose,
+                        constraints: BoxConstraints(minHeight: 200, minWidth: 300),
+                        menuProps: MenuProps(align: MenuAlign.topEnd),
+                      ),
+                    ),
                   ],
                 ),
               ),
@@ -707,7 +744,7 @@ class _dropdownWithGlobalCheckBoxState extends State<_dropdownWithGlobalCheckBox
         onItemAdded: (l, s) => longListCheckBoxValueNotifier.value = _getCheckBoxState(),
         onItemRemoved: (l, s) => longListCheckBoxValueNotifier.value = _getCheckBoxState(),
         onItemsLoaded: (value) => longListCheckBoxValueNotifier.value = _getCheckBoxState(),
-        infiniteScrollProps: InfiniteScrollProps(skip: 0, take: 10),
+        infiniteScrollProps: InfiniteScrollProps(loadProps: LoadProps(skip: 0, take: 10)),
         showSearchBox: true,
         containerBuilder: (ctx, popupWidget) {
           return Container(