+++ lib/src/widgets/popup_menu.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 
 import '../properties/menu_props.dart';
 
@@ -6,7 +7,7 @@ Future<T?> showCustomMenu<T>({
   required BuildContext context,
   required MenuProps menuModeProps,
   required RelativeRect position,
-  required Widget child,
+  required Widget Function(Size size) child,
 }) {
   final NavigatorState navigator = Navigator.of(context);
   return navigator.push(
@@ -48,7 +49,7 @@ class _PopupMenuRouteLayout extends SingleChildLayoutDelegate {
 
     return BoxConstraints.loose(
       Size(
-        constraints.minWidth - position.right - position.left,
+        constraints.minWidth - position.left - position.right,
         constraints.minHeight,
       ),
     ).deflate(
@@ -68,10 +69,14 @@ class _PopupMenuRouteLayout extends SingleChildLayoutDelegate {
     // check if we are in the bottom
     if (y + childSize.height > size.height - keyBoardHeight) {
       y = size.height - childSize.height - keyBoardHeight;
+    } else if (y < 0) {
+      y = 8;
     }
 
     if (x + childSize.width > size.width) {
       x = size.width - childSize.width;
+    } else if (x < 0) {
+      x = 8;
     }
 
     return Offset(x, y);
@@ -87,7 +92,7 @@ class _PopupMenuRoute<T> extends PopupRoute<T> {
   final MenuProps menuModeProps;
   final BuildContext context;
   final RelativeRect position;
-  final Widget child;
+  final Widget Function(Size size) child;
   final CapturedThemes capturedThemes;
 
   _PopupMenuRoute({
@@ -99,9 +104,7 @@ class _PopupMenuRoute<T> extends PopupRoute<T> {
   });
 
   @override
-  Duration get transitionDuration =>
-      menuModeProps.popUpAnimationStyle?.duration ??
-      Duration(milliseconds: 300);
+  Duration get transitionDuration => menuModeProps.popUpAnimationStyle?.duration ?? Duration(milliseconds: 300);
 
   @override
   bool get barrierDismissible => menuModeProps.barrierDismissible;
@@ -113,8 +116,7 @@ class _PopupMenuRoute<T> extends PopupRoute<T> {
   String? get barrierLabel => menuModeProps.barrierLabel;
 
   @override
-  Widget buildPage(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation) {
+  Widget buildPage(BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation) {
     final PopupMenuThemeData popupMenuTheme = PopupMenuTheme.of(context);
     final menu = Material(
       surfaceTintColor: menuModeProps.surfaceTintColor,
@@ -124,11 +126,13 @@ class _PopupMenuRoute<T> extends PopupRoute<T> {
       elevation: menuModeProps.elevation ?? popupMenuTheme.elevation ?? 8.0,
       clipBehavior: menuModeProps.clipBehavior,
       borderRadius: menuModeProps.borderRadius,
-      animationDuration: menuModeProps.popUpAnimationStyle?.duration ??
-          Duration(milliseconds: 300),
+      animationDuration: menuModeProps.popUpAnimationStyle?.duration ?? Duration(milliseconds: 300),
       shadowColor: menuModeProps.shadowColor,
       borderOnForeground: menuModeProps.borderOnForeground,
-      child: child,
+      child: Builder(builder: (ctx) {
+        final menuRenderBox = context.findRenderObject() as RenderBox;
+        return child(menuRenderBox.size);
+      }),
     );
     final MediaQueryData mediaQuery = MediaQuery.of(context);